feat: 결과 확인하기까지 연동
