+++ .eslintrc.json
@@ -1,14 +1,30 @@
 {
-  "rules": {
-    "max-depth": ["error", 2],
-    "max-params": ["error", 3]
-  },
   "env": {
+    "browser": true,
     "es6": true,
-    "node": true
+    "node": false
   },
   "parserOptions": {
-    "ecmaVersion": "latest"
+    "ecmaVersion": "latest",
+    "sourceType": "module"
   },
-  "extends": ["eslint:recommended", "plugin:prettier/recommended"]
+  "extends": ["eslint:recommended"],
+  "plugins": ["import", "prettier"],
+  "rules": {
+    "prefer-const": "error", // 변경되지 않는 변수는 const 사용
+    "no-console": "warn", // console.log 사용 시 경고
+    "eqeqeq": ["error", "always"], // 일치 연산자(===) 사용 필수
+    "no-unused-vars": ["warn", { "args": "none" }], // 사용하지 않는 변수 경고
+    "no-duplicate-imports": "error", // 중복 import 방지
+    "no-shadow": "error", // 변수명이 외부 스코프 변수와 중복 금지
+    "no-multi-spaces": "error", // 여러 개의 공백 금지
+    "comma-dangle": ["error", "never"], // 마지막 쉼표 금지
+    "max-depth": ["error", 2], // 함수 중첩 최대 2단계 제한
+    "max-params": ["error", 3], // 함수의 인자 최대 3개 제한
+    "curly": "error", // 중괄호 사용 강제
+    "consistent-return": "error", // 함수에서 일관되게 return 사용
+    "import/no-unresolved": "error", // 존재하지 않는 모듈 import 방지
+    "import/named": "error", // 존재하지 않는 네임드 import 방지
+    "import/order": ["error", { "groups": ["builtin", "external", "internal"] }]
+  }
 }
