+++ package.json
@@ -18,6 +18,8 @@
     "@types/jest": "^29.5.14",
     "babel-jest": "^29.7.0",
     "eslint": "^9.21.0",
+    "eslint-config-prettier": "^10.0.2",
+    "eslint-plugin-import": "^2.31.0",
     "jest": "^29.6.0",
     "vite": "^6.0.5"
   },
